Organize static assets and extract layout styles
