+++ static/css/layout.css
@@ -1,0 +1,285 @@
+/* ─────────────────────────────────────────────
+   B) Updated mobile tweaks
+   ───────────────────────────────────────────── */
+@media (max-width: 600px) {
+        #appContainer {
+          padding: 10px;
+        }
+
+        #optionsMenu {
+          position: fixed;
+          left: 50%;
+          top: 50%;
+          transform: translate(-50%, -50%);
+        }
+
+        #optionsToggle {
+          display: block;
+        }
+        #chatNotify {
+          display: block;
+        }
+
+      #historyBox {
+        position: fixed;
+        bottom: 0;
+        left: 0;
+        width: 100%;
+        max-height: 0;
+        overflow-y: hidden;
+        border-radius: 0;
+        padding: 0;
+        box-shadow: none;
+        background: var(--bg-color);
+        opacity: 0;
+        transform-origin: bottom center;
+        transform: scale(0);
+        z-index: 50;
+        pointer-events: none;
+      }
+
+      #definitionBox {
+        position: fixed;
+        bottom: 0;
+        right: 0;
+        width: 100%;
+        max-height: 0;
+        overflow-y: hidden;
+        border-radius: 0;
+        padding: 0;
+        box-shadow: none;
+        background: var(--bg-color);
+        opacity: 0;
+        transform-origin: bottom center;
+        transform: scale(0);
+        z-index: 50;
+        pointer-events: none;
+      }
+
+      #chatBox {
+        position: fixed;
+        bottom: 0;
+        left: 0;
+        right: 0;
+        width: 100%;
+        max-height: 0;
+        overflow: hidden;
+        border-radius: 0;
+        padding: 0;
+        box-shadow: none;
+        background: var(--bg-color);
+        opacity: 0;
+        transform-origin: bottom center;
+        transform: scale(0);
+        z-index: 50;
+        display: flex;
+        flex-direction: column;
+        pointer-events: none;
+      }
+
+      body.history-open #historyBox {
+        max-height: 50vh;
+        max-width: calc(98% - 20px);
+        padding: 8px;
+        overflow-y: auto;
+        border-radius: 0;
+        box-shadow: inset 2px 2px 4px var(--shadow-color-dark),
+                    inset -2px -2px 4px var(--shadow-color-light);
+        opacity: 1;
+        transform: scale(1);
+        transition: transform 0.3s ease, opacity 0.3s ease;
+        pointer-events: auto;
+      }
+
+      body.definition-open #definitionBox {
+        display: block;
+        max-height: 40vh;
+        max-width: calc(98% - 20px);
+        padding: 8px;
+        overflow-y: auto;
+        border-radius: 0;
+        box-shadow: inset 2px 2px 4px var(--shadow-color-dark),
+                    inset -2px -2px 4px var(--shadow-color-light);
+        opacity: 1;
+        transform: scale(1);
+        transition: transform 0.3s ease, opacity 0.3s ease;
+        pointer-events: auto;
+      }
+
+      body.chat-open #chatBox {
+        display: flex;
+        max-height: 40vh;
+        max-width: calc(98% - 20px);
+        padding: 36px 8px 8px;
+        overflow: hidden;
+        border-radius: 0;
+        box-shadow: inset 2px 2px 4px var(--shadow-color-dark),
+                    inset -2px -2px 4px var(--shadow-color-light);
+        opacity: 1;
+        transform: scale(1);
+        transition: transform 0.3s ease, opacity 0.3s ease;
+        flex-direction: column;
+        pointer-events: auto;
+      }
+
+      #historyBox h3 {
+        display: none;
+      }
+
+      .history-item .history-guess-tiles .tile {
+        width: 20px;
+        height: 20px;
+        font-size: 12px;
+        line-height: 20px;
+      }
+
+      #board {
+        grid-template-columns: repeat(5, 50px);
+        grid-gap: 5px;
+        max-width: 260px;
+        margin: 0;
+      }
+
+      .tile {
+        width: 50px;
+        height: 50px;
+        font-size: 22px;
+        border-radius: 4px;
+        box-shadow: 3px 3px 6px var(--shadow-color-dark),
+                    -3px -3px 6px var(--shadow-color-light);
+      }
+
+      #guessInput,
+      #submitGuess,
+      #message {
+        display: none;
+      }
+
+      #messagePopup {
+        font-size: 16px;
+      }
+
+      .key {
+        min-width: 32px;
+        height: 48px;
+        margin: 2px;
+        font-size: 16px;
+        padding: 0 5px;
+        box-shadow: 3px 3px 6px var(--shadow-color-dark),
+                    -3px -3px 6px var(--shadow-color-light);
+        border-radius: 6px;
+      }
+
+
+      #resetWrapper {
+        margin-left: 0;
+      }
+      #holdReset {
+        width: 70px;
+        height: 32px;
+        font-size: 14px;
+        border-radius: 8px;
+        box-shadow: 3px 3px 6px var(--shadow-color-dark),
+                    -3px -3px 6px var(--shadow-color-light);
+      }
+      #holdResetProgress {
+        border-radius: 8px;
+      }
+
+      .leaderboard-entry {
+        font-size: 1em;
+        padding: 5px 10px;
+        margin: 0 3px;
+        min-width: 40px;
+      }
+
+      h1 {
+        font-size: 1.5em;
+        margin-bottom: 2px;
+      }
+      #message {
+        font-size: 16px;
+        margin-top: 12px;
+      }
+    }
+
+    /* ─────────────────────────────────────────────
+       C) Medium mode adjustments
+       ───────────────────────────────────────────── */
+    @media (min-width: 601px) and (max-width: 900px) {
+      #historyBox,
+      #definitionBox,
+      #chatBox {
+        width: 220px;
+      }
+
+      #optionsToggle {
+        display: block;
+      }
+
+      body[data-mode='medium'] #historyClose,
+      body[data-mode='medium'] #definitionClose,
+      body[data-mode='medium'] #chatClose {
+        display: block;
+      }
+
+      body[data-mode='medium'] #historyBox,
+      body[data-mode='medium'] #definitionBox,
+      body[data-mode='medium'] #chatBox {
+        position: fixed;
+        top: 50%;
+        left: 50%;
+        transform: translate(-50%, -50%);
+        max-width: 90%;
+        max-height: 80vh;
+        overflow-y: auto;
+        z-index: 80;
+      }
+
+      body[data-mode='medium'] #chatBox {
+        overflow: hidden;
+        display: flex;
+        flex-direction: column;
+      }
+
+      body[data-mode='medium'] #historyBox,
+      body[data-mode='medium'] #definitionBox,
+      body[data-mode='medium'] #chatBox {
+        transform-origin: center center;
+      }
+      body[data-mode='medium']:not(.history-open) #historyBox,
+      body[data-mode='medium']:not(.definition-open) #definitionBox,
+      body[data-mode='medium']:not(.chat-open) #chatBox {
+        transform: translate(-50%, -50%) scale(0);
+        opacity: 0;
+        pointer-events: none;
+      }
+
+      #board {
+        grid-template-columns: repeat(5, 55px);
+        grid-gap: 8px;
+        max-width: 307px;
+      }
+
+      .tile {
+        width: 55px;
+        height: 55px;
+      }
+
+      .key {
+        min-width: 34px;
+        height: 50px;
+      }
+    }
+
+/* Initial layout states */
+#emojiModal,
+#closeCallPopup,
+#infoPopup,
+#optionsMenu {
+  display: none;
+}
+#chatNotify { display: none; }
+#chatForm { display: flex; margin-top: 6px; }
+#chatInput { flex: 1; }
+.key.wide { min-width: 60px; }
